refactor(interferometry): separate files
